Fix the hacked-up PhosphorRender
The return value of a previous call to updatePaintNode is passed back in
as oldNode so it doesn't have to be re-created for every frame. We
static_cast it to the correct type so we can modify its properties. The
previous commit segfaulted because it returned a QSGTransformNode, but
casted oldNode to QSGGeometryNode. The QSGGeometryNode is a child of the
QSGTransformNode, and must be obtained from the parent node and then
casted.

This also cleans up the vestiges of having the transformation matrix as
part of the Material. We should probably take this approach of using
QSGTransformNode instead of stuffing transformation into the custom
material even in versions that retain the custom material.