main.cpp: don't add the current directory to the library path
This messes up OS X builds since it causes the system to search for the
qt platform plugins in the current working directory before the correct
path that's specified in qt.conf.

In essence, this should be unnecessary and every platform should use a
proper qt.conf to specify various paths.