First working implementation of firmware update
The device list shows which devices are present and which devices need a
firmware update. Besides this it shows if there is a device in programming
mode is also present.
User decides which device to have its FW updated.
The FW of a device cannot be updated if there is another device present
which is set in programming mode. The latter shall be programmed or
disconnected before continuing to program the former device.

Signed-off-by: Dan Nechita <Dan.Nechita@analog.com>