windows: Store firmware file to a path more suited for persistent data
The 'Roaming' folder under AppData is the new location where the
downloaded firmware is being stored.
This commit fixes the firmware update not working on windows 10 due to
the fact that the firmware file was not being stored in the same folder as
the executable making the FW file inaccessible. This happens when running
the application through its shortcut.

Signed-off-by: Dan Nechita <Dan.Nechita@analog.com>