QMake: Assume we use the Unix date utility instead the windows one
On windows the date utility is quite different from the Unix one.
To get date on windows, the following should be used: 'date /t' instead of
'date'(Unix), otherwise the shell from which the command was called will
block waiting for the user to input a new date.
The windows date utility returns the date in a local format, making it
hard to guess which is the month and which is the day. A standard format
is preferred instead. Therefore assume we only use the Unix date utility.

This means that if building on a windows machine, the unix utility should
be installed and the shell should be configured to use the Unix utility.

Signed-off-by: Dan Nechita <Dan.Nechita@analog.com>