QMake: Get date when building and only when on unix systems
The execution of system(<command>) occurs when qmake parses the project
file. It is best to get the date and commit during the actual build.
Therefore 'system(<command>)' gets replaced with 'shell <command>'
which will be copied to the Makefile where will get executed during the
build.

However adding shell commands won't work if the makefile is parsed by
nmake or jom, which happens when qmake creates a msvc project and that
projects gets build by msbuild. Therefore, skip getting the date on
windows but configure the build machine to get the date and supply it to
the qmake during the creating of the msvc project.

Signed-off-by: Dan <Dan.Nechita@analog.com>