Delayed Acquisition: Apply delay only when changing the buffer size
This avoids situations where the component that display data uses the new delay value that user just entered but the data buffer did not get the change to be resized accordingly.
E.g. having a 10k buffer and a initial delay of 0 samples, if delay is set to 8k, it is needed to first request a buffer 18k samples before the component that display data to use the 8k value to position itself inside the buffer. Otherwise segfault occurs.

Signed-off-by: Dan Nechita <Dan.Nechita@analog.com>