Logging data on another thread.
For performance reasons, we moved the logging process on another thread.
We are now not anymore requesting the logger to log every single sample every time it is generated,
but requesting to log a larger amount of samples at a time, on another thread.
Therefore, Pixelpulse is not waiting anymore for the logger to end it's logging process.

Signed-off-by: Cristi Iacob <cristian.iacob@analog.com>