Resolved the delay which appeared in multi-device mode.
There was a delay when plotting the points. It appeared only in multi-device mode.
(for example, with 2 connected devices, it took ~12s to fill the plot for the sample time of 10s and the current/voltage variations were plotted a little later than they really appeared).
This is because the number of graphical updates to be made was pretty big.
OpenGL works faster when it receives a larger amount of data to plot, because it reduces the number of calls to the GPU, which are expensive.
Therefore, I changed the number of samples requested to libsmu from 1000 to 10000, in order to send them all at a time to OpenGL and make it plot them faster.

Signed-off-by: Cristi Iacob <cristian.iacob@analog.com>